+++ EMS/src/TimeSheet.Modules.EmploymentManagement.Domain/Security/RoleCodes.cs
@@ -3,5 +3,7 @@ namespace TimeSheet.Modules.EmploymentManagement.Domain.Security;
 public static class RoleCodes
 {
     public const string Admin = "Admin";
-    public const string Basic = "Basic";
+    public const string HR = "HR";
+    public const string Manager = "Manager";
+    public const string Developer = "Developer";
 }